Split code up, embed markdown template, add beginnings of sections
