+++ octopus2/src/docparser/markdown.md
@@ -4,12 +4,24 @@
 {{end}}
 {{- $path := .Filepath -}}
 
-{{if .Consts -}}
+## Table of Contents
+
+{{range .Sections}}{{if .Name}}- {{ .Name }}
+{{end}}{{if .Consts}}  - Constants
+{{range .Consts}}    - [{{.Name}}](#{{.Name}})
+{{end}}{{end}}{{if .Macros}}  - Macros
+{{range .Macros}}    - [{{.Name}}](#{{.Name}})
+{{end}}{{end}}{{if .Routines}}  - Routines
+{{range .Routines}}    - [{{.Name}}](#{{.Name}}){{end}}{{end}}{{end}}
+
+{{range .Sections}}{{if .Name}}# {{ .Name }}
+
+{{end -}}{{if .Consts -}}
 ## Constants
 {{end -}}
 
 {{range .Consts}}
-### {{ .Name }}
+### `{{ .Name }}`
 
 _{{ $path }}:{{ .Line }}_
 {{if .Description}}
@@ -23,7 +35,7 @@ Value: `{{ .Value }}`
 {{end -}}
 
 {{range .Macros}}
-### {{ .Name }}
+### `{{ .Name }}`
 
 _{{ $path }}:{{ .Line }}_
 {{if .Description}}
@@ -36,14 +48,15 @@ _{{ $path }}:{{ .Line }}_
 {{- end}}
 
 {{if .Routines -}}
-## Routines
+### Routines
 {{end -}}
 
 {{range .Routines}}
-### {{ .Name }}
+### `{{ .Name }}`
 
 _{{ $path }}:{{ .Line }}_
 {{if .Description}}
 {{ .Description }}
+{{- end}}
 {{- end}}
 {{- end}}
